Move ShipData to ship folder
